Move kerberos to its own sub module
