+++ airflow/security/kerberos.py
