+++ airflow/security/utils.py
